refactor(UserController): CQRS 적용하기
